Fixed Xbox 360 EULA Freeze
Xbox 360 backwards compatibility would freeze when accepting the EULA in game as a result of patching the "Do Not Agree" code that would exit to dashboard. The issue being that the patched instructions would leave bytes from the original instructions over, and the emulator seems to have some issue with this. Possible prefetching instructions for optimization. This would cause freezing even though that code path wasn't actually executed.